+++ examples/official-site/colors.sql
@@ -1,0 +1,4 @@
+select 'dynamic' as component, properties FROM example WHERE component = 'shell' LIMIT 1;
+
+select 'list' as component,
+    'Colors' as title;